+++ src/Radiance_Monitor/image_gen/ush/plot_summary.sh
@@ -6,11 +6,9 @@
 #
 #------------------------------------------------------------------
 
-set -ax
-export list=$listvars
 SATYPE2=$SATYPE
 
-echo "Start plot_summary.sh"
+echo; echo "Start plot_summary.sh"
 
 #------------------------------------------------------------------
 # Set environment variables.
@@ -51,49 +49,30 @@ for type in ${SATYPE2}; do
    #
    while [[ $cdate -le $edate ]]; do
 
-      if [[ $REGIONAL_RR -eq 1 ]]; then
-         tdate=`$NDATE +6 $cdate`
-         day=`echo $tdate | cut -c1-8`
-         cyc=`echo $cdate | cut -c9-10`
-         . ${IG_SCRIPTS}/rr_set_tz.sh $cyc
-      else
-         day=`echo $cdate | cut -c1-8`
-         cyc=`echo $cdate | cut -c9-10`
-      fi
-
-      #----------------------------------------------------
-      #  Attempt to locate the extracted ieee data files.
-      #
-      ieee_src=${TANKverf}/${RUN}.${day}/${cyc}/${MONITOR}
-      if [[ ! -d ${ieee_src} ]]; then
-         ieee_src=${TANKverf}/${RUN}.${day}/${MONITOR}
-      fi
-      if [[ ! -d ${ieee_src} ]]; then
-         ieee_src=${TANKverf}/${MONITOR}.${day}
-      fi
-      if [[ ! -d ${ieee_src} ]]; then
-         ieee_src=${TANKverf}/${RUN}.${day}
-      fi
-
       #-----------------------------------------------------------
       #  Locate the data files, first checking for a tar file,
       #  and copy them locally.
       #
-      if [[ -e ${ieee_src}/radmon_time.tar && -e ${ieee_src}/radmon_time.tar.${Z} ]]; then
-         echo "Located both radmon_time.tar and radmon_time.tar.${Z} in ${ieee_src}.  Unable to plot."
-         exit 23
-				      
-      elif [[ -e ${ieee_src}/radmon_time.tar || -e ${ieee_src}/radmon_time.tar.${Z} ]]; then
-         files=`tar -tf ${ieee_src}/radmon_time.tar* | grep ${type} | grep ieee_d`
-	 if [[ ${files} != "" ]]; then 
-            tar -xf ${ieee_src}/radmon_time.tar* ${files}
-         fi
+      ieee_src=`$MON_USH/get_stats_path.sh --run $RUN --pdate ${cdate} \
+	        --net ${RADMON_SUFFIX} --tank ${R_TANKDIR} --mon radmon`
 
-      else				
-         files=`ls ${ieee_src}/time.*${type}*ieee_d*`
-         for f in ${files}; do
-            $NCP ${f} .
-         done
+      if [[ -d ${ieee_src} ]]; then
+         if [[ -e ${ieee_src}/radmon_time.tar && -e ${ieee_src}/radmon_time.tar.${Z} ]]; then
+            echo "Located both radmon_time.tar and radmon_time.tar.${Z} in ${ieee_src}.  Unable to plot."
+            exit 23
+				      
+         elif [[ -e ${ieee_src}/radmon_time.tar || -e ${ieee_src}/radmon_time.tar.${Z} ]]; then
+            files=`tar -tf ${ieee_src}/radmon_time.tar* | grep ${type} | grep ieee_d`
+       	    if [[ ${files} != "" ]]; then 
+               tar -xf ${ieee_src}/radmon_time.tar* ${files}
+            fi
+
+         else				
+            files=`ls ${ieee_src}/time.*${type}*ieee_d*`
+            for f in ${files}; do
+               $NCP ${f} .
+            done
+         fi
       fi
 
       adate=`$NDATE +${CYCLE_INTERVAL} ${cdate}`
@@ -106,7 +85,7 @@ for type in ${SATYPE2}; do
    #  Remove 'time.' from the *ieee_d file names.
    #
    prefix="time."
-   dfiles=`ls *.ieee_d`
+   dfiles=`ls *.ieee_d 2>/dev/null`
    if [[ $dfiles = "" ]]; then
       echo "NO DATA available for $type, aborting summary plot"
       continue
@@ -156,9 +135,10 @@ EOF
    #       e) copy the [satype].sum.txt file to $TANKverf/imgn/{suffix}/pngs/summary/.
    #       f) clean up
 
-   echo "BEGIN data file generation:"
-   rm -f $timesf
-   
+   if [[ -e $timesf ]]; then
+      rm -f $timesf
+   fi
+
    if [[ ! -s summary.x ]]; then
       $NCP ${IG_EXEC}/radmon_ig_summary.x summary.x 
    fi
@@ -167,15 +147,19 @@ EOF
    tac tmp > $timesf
    rm tmp
 
-   rm -f $usef
-   cat ./${type}.ctl | grep iuse | gawk '{print $8}' > $usef
+   if [[ -e ${usef} ]]; then
+      rm -f ${usef}
+   fi
+   cat ./${type}.ctl | grep iuse | gawk '{print $8}' > ${usef}
 
    nchanl=`cat ./${type}.ctl | grep title |gawk '{print $4}'`
 
    #------------------------------
    #  build chan.txt using ctl file
    #
-   rm $chanf
+   if [[ -e ${chanf} ]]; then
+      rm ${chanf}
+   fi
    grep iuse ${type}.ctl | gawk '{print $5}' > $chanf
 
    ncycle=`cat $timesf | wc -l`
@@ -203,7 +187,6 @@ EOF
    rm -f input
    cp ${input} input
    ./summary.x < input > out.${type}
-   echo "END data file generation:"
 
    rm -f ${input}
    rm -f ${type}*ieee_d
@@ -234,7 +217,7 @@ cd $tmpdir
 cd ../
 rm -rf $tmpdir
 
-echo "End plot_summary.sh"
+echo "End plot_summary.sh"; echo
 
 exit
